refactor : 사정사 검수 화면 로컬 컴포넌트 파일 분리
* 초안 패널의 본문·안내·스켈레톤을 형제 파일로 분리
* 데스크톱 목록의 결과 영역과 두 스켈레톤을 형제 파일로 분리
* 첨부 자료의 미리보기 모달과 파일 타입 아이콘을 형제 파일로 분리
* 이미지 판별 함수를 _model로 이동
* 확정 보상액 입력을 형제 파일로 분리
