+++ apps/web/src/app/partner/review/[reportId]/_components/AttachmentSection.tsx
@@ -1,75 +1,18 @@
 "use client";
 
-import { useEffect, useRef, useState } from "react";
+import { useState } from "react";
 import { cn } from "@/shared/lib/utils";
 import { Button, buttonVariants } from "@/shared/ui/Button";
-import { FileText } from "@/shared/ui/icons/FileText";
-import { ImageIcon } from "@/shared/ui/icons/ImageIcon";
 import { Search } from "@/shared/ui/icons/Search";
-import { X } from "@/shared/ui/icons/X";
-import { useFocusTrap } from "@/shared/lib/use-focus-trap";
 import type { ReviewAttachment } from "../_model/types";
-
-function isImageMime(mimeType: string): boolean {
-  return /^image\//i.test(mimeType) || /jpe?g|png|gif|webp/i.test(mimeType);
-}
+import { AttachmentPreviewModal } from "./AttachmentPreviewModal";
+import { FileTypeIcon } from "./FileTypeIcon";
 
 /** "application/pdf" → "PDF", "image/jpeg" → "JPG". */
 function mimeLabel(mimeType: string): string {
   const subtype = mimeType.split("/")[1] ?? mimeType;
   if (/jpe?g/i.test(subtype)) return "JPG";
   return subtype.toUpperCase();
-}
-
-function PreviewModal({ file, onClose }: { file: ReviewAttachment; onClose: () => void }) {
-  const isImage = isImageMime(file.mimeType);
-  const panelRef = useRef<HTMLDivElement>(null);
-  useFocusTrap(panelRef, true);
-
-  useEffect(() => {
-    const onKey = (e: KeyboardEvent) => e.key === "Escape" && onClose();
-    document.addEventListener("keydown", onKey);
-    return () => document.removeEventListener("keydown", onKey);
-  }, [onClose]);
-
-  return (
-    <div
-      ref={panelRef}
-      role="dialog"
-      aria-modal="true"
-      aria-label={`${file.name} 원본 미리보기`}
-      onClick={onClose}
-      className="fixed inset-0 z-50 flex flex-col items-center justify-center gap-3 bg-ink/60 p-4 sm:p-8"
-    >
-      <div className="flex w-full max-w-4xl items-center justify-between gap-2 text-white">
-        <p className="text-[0.9375rem] font-semibold">{file.name}</p>
-        <button
-          type="button"
-          onClick={onClose}
-          aria-label="닫기"
-          className="rounded-full p-2 transition hover:bg-white/15"
-        >
-          <X className="size-[1.375rem]" />
-        </button>
-      </div>
-      <div
-        onClick={(e) => e.stopPropagation()}
-        className="flex max-h-[80vh] w-full max-w-4xl items-center justify-center overflow-auto rounded-card bg-card"
-      >
-        {isImage ? (
-          // eslint-disable-next-line @next/next/no-img-element
-          <img src={file.url} alt={file.name} className="max-h-[80vh] w-auto object-contain" />
-        ) : (
-          <iframe
-            src={file.url}
-            title={file.name}
-            sandbox=""
-            className="h-[80vh] w-full"
-          />
-        )}
-      </div>
-    </div>
-  );
 }
 
 export interface AttachmentSectionProps {
@@ -79,14 +22,6 @@ export interface AttachmentSectionProps {
 function fileMeta(file: ReviewAttachment): string {
   const pages = file.pageCount != null ? ` · ${file.pageCount}page` : "";
   return `${mimeLabel(file.mimeType)}${pages}`;
-}
-
-function FileTypeIcon({ mimeType }: { mimeType: string }) {
-  return isImageMime(mimeType) ? (
-    <ImageIcon className="size-[1.125rem]" />
-  ) : (
-    <FileText className="size-[1.125rem]" />
-  );
 }
 
 function downloadAll(attachments: ReviewAttachment[]) {
@@ -212,7 +147,7 @@ export function AttachmentSection({ attachments }: AttachmentSectionProps) {
       </div>
 
       {previewOpen && selected && (
-        <PreviewModal file={selected} onClose={() => setPreviewOpen(false)} />
+        <AttachmentPreviewModal file={selected} onClose={() => setPreviewOpen(false)} />
       )}
     </div>
   );
