+++ apps/web/src/app/partner/review/[reportId]/_model/mime.ts
@@ -1,0 +1,3 @@
+export function isImageMime(mimeType: string): boolean {
+  return /^image\//i.test(mimeType) || /jpe?g|png|gif|webp/i.test(mimeType);
+}
