+++ apps/web/src/app/partner/review/_components/DraftPlaceholder.tsx
@@ -1,0 +1,7 @@
+export function DraftPlaceholder() {
+  return (
+    <div className="rounded-card-lg border border-dashed border-line bg-paper-2 p-8 text-center text-sm text-ink-3">
+      사건을 선택하면 AI 초안을 확인할 수 있어요.
+    </div>
+  );
+}
