+++ apps/web/src/app/partner/review/_components/DraftSkeleton.tsx
@@ -1,0 +1,8 @@
+export function DraftSkeleton() {
+  return (
+    <div className="space-y-3">
+      <div className="h-44 animate-pulse rounded-card-lg bg-line-2" />
+      <div className="h-52 animate-pulse rounded-card-lg bg-line-2" />
+    </div>
+  );
+}
